Added File.Exists check to DeleteCaches
An oversight, but the File.Exists check was not included until now.

+++ Model/RMModel.cs
@@ -1350,7 +1350,7 @@ namespace RhinoMobile.Model
 		{
 			DeleteCaches ();
 
-			if (DocumentsFilename != string.Empty)
+			if (DocumentsFilename != string.Empty && File.Exists(ModelPath))
 				System.IO.File.Delete(ModelPath);
 
 			DocumentsFilename = string.Empty;
